storage,coord: remove the concept of builtin logs from storage
Materialize presents timely/differential logging relations as built-in
sources. They are completely unlike sources managed by the storage
layer, though, as they only exist within the context of a compute
replica that has introspection enabled.

This removes all knowledge of these logging relations from the storage
layer. Instead, these relations are represented explicitly in the
adapter layer as a new `BuiltinLog` catalog item type. This makes the
special cases for these logs far more explicit; previously those special
cases were largely the result of emergent behavior, leading to bugs like
 #12644.

When exposed to the user, we continue to call these logging relations
"sources", to avoid introducing yet another noun. For example, logging
relations continue to appear in the output of `SHOW SOURCES` and are
listed as an object of type `source` in `SHOW OBJECTS`. But we're now
well positioned to revisit this decision in the future, if we like.

Fix #12644.