persist: also retry BlobMulti and Consensus init
I initially didn't do this so that configuration errors would make their
way back up to the caller. However, this intuition is a bit of a
holdover from the old world. When we always control the deployment,
we're free to retry even initialization of external systems. Logs and
other introspection will show what's going on if we do ever have a
misconfiguration (plus perhaps they're less likely because we just have
to get them right once).

Immediate motivation is that Brennan saw a transient s3 failure at
startup and saw a process crash as a result.