sql: temporarily allow secret references in with option values
To facilitate the connector/secret work, temporarily allow secret
references in WITH option clauses, as in:

    CREATE SOURCE ... WITH (sasl_password = SECRET a.b.c);

This commit only wires up the parser support for these references. The
planner will reject any references of this form.

The key trick here is that secret references are represented as

    WithOptionValue::Secret(T::ObjectName)

rather than:

    WithOptionValue::ObjectName(UnresolvedObjectName)

That means they will participate in name resolution like normal, so that
a `CreateSourceStatement<Aug>` will contain the global ID for all
embedded secret references.

Once we mandate the use of connectors with explicit fields for secret
references, we can remove this.