sql: always resolve all names as first step of planning
This does for the plan_* suite of functions what #12830 did for the
describe_* suite of functions. The `sql::plan` function now
unconditionally performs name resolution as the first step of planning.

A few statements need to opt some of their names out of the forced name
resolution. `DROP` and `ALTER` both need to suppress name resolution
errors if the `IF EXISTS` clause is present. `PREPARE` and `DECLARE`
need to defer name resolution on their embedded statements. The approach
taken is to make these names non-generic in the AST; e.g., where you'd
normally use a `T::ObjectName`, you use an `UnresolvedObjectName`
instead; this leaves resolving the names to the relevant plan_ function.
In the future we might want to investigate alternatives, like definining
`Aug::ObjectName` as `Result<ResolvedObjectName, anyhow::Error>`.

This is further work towards #5302.