coord,sql: split name resolution into a separate pass
This commit makes the coordinator responsible for performing name
resolution. It does so after parsing, but before purification and
planning. The correct order is enforced by the type system, as the
relevant functions in the SQL planner now deal only in <Aug> statements
rather than <Raw> statements.

This commit also removes the undocumented and untested `CREATE VIEWS
(...)` variant that permitted creating multiple views in one shot, as it
is incompatible with the new model of name resolution, as we no longer
have granular tracking of dependencies *within* a single statement.  But
that seems fine, as the statement seems dubiously useful anyway.
(`CREATE VIEWS FROM SOURCE` remains untouched.)

Fix #5302.