persist: take a bytes::Bytes for value in Blob::set
The production impl of this (s3) was already doing this conversion.
Moving it up a level to the Blob::set allows us to do a (trivial)
`Bytes::clone` instead of `Vec<u8>::clone`. Note that this vec clone was
happening even on the first attempt because we had to keep the original
value around in case we needed to retry. These values are large, so this
is probably significant.