persist: add initial attempt at monitoring metrics
This certainly contains some we won't end up needing and also doesn't
contain some we will, but gotta start somewhere.

Old persist didn't take advantage of the CounterVec stuff because we
wanted to make it obvious that nothing was being leaked to third party
metrics. Third party metrics are no longer a thing, so take advantage of
prometheus labels to keep down the maintenance burden this time.