orchestrator: introduce service watcher
This commit introduces the `watch_services` method to the
`NamespacedOrchestrator` trait, and corresponding implementations to the
concrete orchestrators. `watch_services` supports continuously listening
for status changes of services in the respective namespace.

The `ServiceStatus` enum is kept simple in the initial implementations,
services can only be ready or not ready. This will probably become more
fine-grained in the future.

The kubernetes orchestrator regards a service as ready if its "Ready"
condition is fulfilled, and not ready otherwise. The process
orchestrator simply considers every running process as ready.