persist: flush out update batches to s3 to bound memory use
Previously, a BatchBuilder would buffer all updates in memory and flush
them all out as a single BlobTraceBatchPart when the persist user called
BatchBuilder::finish. Now, whenever a BatchBuilder::add call causes
ColumnarRecordsVecBuilder to finish a chunk of updates, they are sent
off in a task to be encoded and uploaded to blob storage. If too many of
these are outstanding (perhaps s3 is down or slow), BatchBuilder::add is
backpressured until the number that are outstanding is back under the
threshold.