compute: Ignore invalid protocol prefixes
Implement a stop-gap to protect the computed against malformed prefixes
of the compute command protocol. This ignores all messages until the first
CreateInstance command. This should be correct because the invalid prefix
of messages should also be folded into the command history, which drives
the hydration of each replica. Ignoring the prefix is fine because the
intention of the messages will be picked up later.

Related: #12233

Signed-off-by: Moritz Hoffmann <mh@materialize.com>