coord,materialized: add test for purification race
Add a test for connectors that go missing during purification due to a
concurrent `DROP CONNECTION`. The existing code in the coordinator
successfully prevented a crash but produced a confusing error message
about cached plans; this commit modifies that code a bit to produce a
better error message about the missing connection.

Fix #12983.