stash: Prevent ID reuse for compute instances (#13110)
The default behavior for ID selection in stash collections is to take
the maximum of all current IDs and add one. This means that after
deleting an item in a stash collection, it's ID can be reused, if the
collection is using the default behavior.

This is problematic for compute instances due to how TAIL is
implemented. TAIL creates a temporary sink that stores the ID of the
compute instance that it was created in. When the TAIL is canceled,
the associated sink is deleted from the compute instance, if the
compute instance still exists. The compute instance ID is used to
identify if the compute instance exists. The TAIL can outlive the
compute instance if the compute instance is deleted while the TAIL is
running. If the compute instance is deleted and new one is created,
then the new compute instance may share the same ID as the old one. The
TAIL will then mistakenly try and delete the sink from this new compute
instance and cause a panic.

This commit changes ID selection for compute instances so that they are
always monotonically increasing and no ID is ever reused. The
implementation is based on GlobalId allocation.

Fixes #13100