storage: fix reclocking correctness bug
The reclock operator previously exposed an API that allowed the caller
to map the max observed offset to a reclocked timestamp. This API
allowed for non-definite behaviour in the face of restarts.

The way this would happen is that on the first run a `storaged` would
read data until some offset `X`, and produce some bindings, and then
upon restart another `storaged` instance would read data until some
offset `Y > X` and also try to produce bindings. Since `Y > X` a later
timestamp would be chosen but instead of reclocking only the messages
at offsets `X < offset <= Y` at this new timestamp all the data was
being pushed to the future.

The result is that duplicated data end up in the output.

This PR fixes the above by changing the relock operator API to reclock
individual messages intead of just the upper frontier and also breaks
down the various steps it performs in order to aid understanding and
allow unit testing the code.