storage: resume ingestion on restart
This is the first part of making source ingestion fault tolerant. With
this PR the storage controller durably records the persist shard ids of
all sources and the `storaged` process is changed in such a way that it
can resume appending data to the definite collection.

Tables are an exception to the above and maintain the previous behaviour
where new shards are minted every time the storage controller restarts
and are never durably persisted.

Postgres sources are also an exception to the above because unlike
sources like Kafka we lose the ability to generate a full snapshot at
the LSN of a replication slot once the initial transaction that created
it has ended. This means that a `storaged` instance that just restarted
has no way of reproducing the previous stream. This won't be an issue
when we wire up start offsets correctly in which case the postgres
source reader will completely skip the snapshot phase and just read from
the replication slot.

Lastly, while this PR makes source ingestion fault tolerant in the sense
that after restarts we can resume correctly it does not yet do the
optimization that we skip ahead in the upstream system. For example for
Kafka we will re-read everything from the beginning and just drop all
data that is not beyond the current upper frontier on the floor. This is
wasteful but has the sideeffect of fully re-hydrating the in-memory
state of the ingestion pipeline.

After this gets merged I'll work on a follow up PR that implements the
optimization described above for simple cases like NONE envelope
sources.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>