orchestrator-process: only emit new status events
Prior to this commit, the process orchestrator's status watcher would
emit a status event for every existing process every 5 seconds. This is
now changed to only emit events for processes we have not already
emitted events for. The purpose is reducing some noise that complicates
debugging of performance problems in Materialize.