coord: Fix read-then-write serializability issue (#13336)
Previously, read-then-write statements would release the global write
lock, before queuing their writes for a group commit. This allowed two
read-then-write statements to interleave their reads and writes, which
can lead to a serializable consistency violation.

An example of one such violation is the following:
1. Txn 1 acquires lock
2. Txn 1 reads from A
3. Txn 1 releases lock
4. Txn 1 submites write to A
5. Txn 2 acquires lock
6. Txn 2 reads from A
7. Txn 2 releases lock
8. Txn 2 submits write to A
9. Group commit executes
There is no serial order that would be equivalent to this order.

This commit updates the behavior of read-then-write statements to hold
on to the global write lock until group commit executes. This means
that no two read-then-write statements can be part of the same group
commit.

However, we can include one read-then-write statements and many write
transactions in the same group commit without violating
serializability. Specifically, the serial order would be to order all
of the write transactions after the read-then-write statement.