storage: use WriteHandle::append when sinking data
There is no good reason to use `compare_and_append` while sinking data
and switching to `append` has the added benefit of giving us a quick and
easy HA option for ingestion.

The original code did use `append` and I incorrectly changed it to
`compare_and_append` thinking that the final behaviour would be better.

After discussing with @aljoscha we concluded that `append` is the right
thing to do, at least until we start seeing performance issues with
persist.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>