storage: reorganize types to reflect current state of affairs
This PR reorganizes the types to better reflect the current state of
affairs in how storage and materialize in general behaves. Previously
there were many places that code was leading to `unreachable!()`
statements because of that.

In the high level the following changes are performed:

In the `StorageController` we separate the concept of a storage
collection from an ingestion. Consequently the
`StorageController::create_sources` method has been renamed into
`create_collections`. Each collection is described by a
`CollectionDescription` (previously named `CreateSourceRequest` which is
responsible for telling storage what this collection should be.
Importantly, the `CollectionDescription` is *not* forced to have any
ingestion attached to it. This is reflected in the type by having its
`ingestion` field be optional. This naturaly splits tables, which are
handled entirely by the `StorageController` and have `ingestion: None`
from ingested sources which have the field populated with the
appropriate data.

Based on the above the types that describe are also massively
simplified. The `SourceConnection` enum that previously had `External`
and `Local` variants now is a struct that contains what `External`
contained. There is no need for describing the `Local` variant anymore.
Consequently `ExternalSourceConnection` is renamed to simply
`SourceConnection`.

Storage types were previously carrying around timeline information but
this is a purely ADAPTER concept. Therefore the timeline fields have
been removed from storage collections and are now maintained by the
coordinator for each object.

Another cleanup that this PR does is that is reduces the amount of
information exposed to COMPUTE when a source is imported. This is
reflected by the changes in `SourceInstanceDesc` which previously
carried the whole ingestion description just to be able to reason about
monotonicity. This is now removed and monotonicy information is
calculated in the planner and passed as a value in the
`DataflowDescription`, just like the rest of the objects in it.

Finally, old code paths that checked for single materialization sources
and the respected plumbing of errors in the coordinator have been
removed as there is no longer a concept of a source instance.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>