proto: split common proto stuff into a new crate
This is mostly for dependency reasons. We want a RustType impl for
ShardId and for rust trait coherency it need to next to either RustType
or ShardId.

Persist doesn't want to depend on repr (a) for reasons of abstraction
purity and (b) for dependency reasons. It's also not great for repr to
depend on persist because then a small change to persist causes
./bin/environmentd to recompile quite of bit of the codebase (why I'm
doing this now).

Both of these are avoided by splitting proto out into a new
low-dependency crate. It also feels correct for the proto library code
to have its own crate.