service: prepare for epoch support in GrpcServer
This commit reworks the concurrency setup in `GrpcServer`. The new
implementation directly uses the task created by the gRPC server to
forward commands to the underlying client, with a straightforward
cancellation protocol to fence out a pre-existing client.

This results in an implementation whose correctness is a bit easier (for
me) to reason about, and leaves a natural place to slot in epoch
checking once we tackle #13377.

As part of this work, remove the `--linger` and `--reconcile` options.
These are required for correct functioning in Materialize Cloud, so
better not to have it optional locally, IMO. The `--reset` option to
`bin/environmentd` is updated to find and kill any orphan `storaged` and
`computed` processes, which should help mitigate any bugs in the
reconciliation logic.

Touches #13377.